3.3. Вычисление площади треугольника. + 3.4. Максимум из трех чисел. Add classes Triangle,TriangleTest and edit Max, MaxTest - add method maxOtThree.

+++ chapter_001/src/main/java/ru/frechman/condition/Point.java
@@ -5,8 +5,8 @@ package ru.frechman.condition;
  */
 public class Point {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     /**
      * Задаёт координаты точки.
@@ -14,7 +14,7 @@ public class Point {
      * @param x координата по оси X (абсцисса).
      * @param y координата по оси Y (ордината).
      */
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
